% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLogger.R
\docType{data}
\name{log.colors}
\alias{log.colors}
\title{Default Logging Colors}
\format{An object of class \code{character} of length 5.}
\usage{
log.colors
}
\description{
Default colors to use as they correspond to the log.levels
}
\keyword{datasets}
