% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLogger.R
\docType{data}
\name{log.levels}
\alias{log.levels}
\title{Default Log Levels}
\format{An object of class \code{character} of length 5.}
\usage{
log.levels
}
\description{
Log levels that are used to determine which messages to print
}
\keyword{datasets}
