% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{Logger}
\alias{Logger}
\title{Shiny Logger}
\usage{
Logger(threshold = log.levels[5], colors = log.colors)
}
\arguments{
\item{threshold}{Level at which to cut logging: See log.levels}

\item{colors}{Colors to use for logging, if the crayon package is available}
}
\value{
List of class type `logging`
}
\description{
Main function for generating a logger to use within Shiny applications
}
\examples{
logger = Logger("INFO")
logger$error("Show error")
logger$warn("Show warning")
logger$info("Show info")
logger$debug("Show debug")
logger$trace("Show trace")
options("shiny.logLevel" = "TRACE")
logger$trace("Show trace")
set.threshold("WARN")
logger$info("Show info")
logger$warn("Show warning")

}
