% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{print.log}
\alias{print.log}
\title{S3 print for log messages}
\usage{
\method{print}{log}(x, ...)
}
\arguments{
\item{x}{object of class log}

\item{...}{optional arguments to print (ignored)}
}
\description{
S3 print for log messages
}
