% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.threshold.R
\name{set.threshold}
\alias{set.threshold}
\title{Set Logging Threshold}
\usage{
set.threshold(threshold = "DEBUG")
}
\arguments{
\item{threshold}{Default value to set as the threshold}
}
\description{
Set the level for logging messages
}
\details{
Used to set the level at which messages are printed out.  If set to INFO, messages of DEBUG and TRACE will not be printed.
}
